soc/intel/skylake: Fix compilation under x86_64
Reviewed-on: https://review.coreboot.org/c/coreboot/+/48177
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/intel/skylake/acpi.c
@@ -169,7 +169,7 @@ void acpi_create_gnvs(struct global_nvs *gnvs)
 
 #if CONFIG(CONSOLE_CBMEM)
 	/* Update the mem console pointer. */
-	gnvs->cbmc = (u32)cbmem_find(CBMEM_ID_CONSOLE);
+	gnvs->cbmc = (u32)(uintptr_t)cbmem_find(CBMEM_ID_CONSOLE);
 #endif
 
 	if (CONFIG(CHROMEOS)) {
@@ -554,7 +554,7 @@ void southbridge_inject_dsdt(const struct device *device)
 
 		/* Add it to DSDT.  */
 		acpigen_write_scope("\\");
-		acpigen_write_name_dword("NVSA", (u32) gnvs);
+		acpigen_write_name_dword("NVSA", (u32) (uintptr_t)gnvs);
 		acpigen_pop_len();
 	}
 }
